requirements.txt
gunicorn
falcon
pony
psycopg2cffi
python-memcached